+++ packages/orq-rc/RELEASES.md
@@ -1939,3 +1939,13 @@ Based on:
 - [typescript v3.7.0-rc.8] packages/orq-rc
 ### Releases
 - [NPM v3.7.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/3.7.0-rc.8 - packages/orq-rc
+
+## 2025-06-03 21:52:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.555.1 (2.618.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.7.0-rc.9] packages/orq-rc
+### Releases
+- [NPM v3.7.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/3.7.0-rc.9 - packages/orq-rc